+++ src/dex/chains/mod.rs
@@ -1,6 +1,8 @@
 pub mod chain;
+pub mod gas;
 pub mod tokens;
 
 // Re-export
 pub use chain::ChainId;
+pub use gas::default_gass_fee_usd;
 pub use tokens::Token;